feat: add windows service features
