+++ go.mod
@@ -7,4 +7,7 @@ require (
 	gopkg.in/yaml.v3 v3.0.1
 )
 
-require golang.org/x/sys v0.13.0 // indirect
+require (
+	github.com/kardianos/service v1.2.2 // indirect
+	golang.org/x/sys v0.13.0 // indirect
+)
