+++ install-as-windows-service.ps1
@@ -1,0 +1,2 @@
+$CurrentPath = $PWD.Path
+New-Service -Name "GoAutoBackup" -BinaryPathName $CurrentPath\go-auto-file-backup.exe
