+++ uninstall-as-windows-service.ps1
@@ -1,0 +1,1 @@
+Remove-Service -Name "GoAutoBackup"
